Refactor get_connection_id
