+++ api/watch-game-disconnect.py
@@ -4,7 +4,7 @@ import json
 import decimal
 import uuid
 from shared.response import * 
-from shared.api_gateway import parse_event
+from shared.api_gateway import parse_event, get_connection_id
 
 
 dynamodb = boto3.resource('dynamodb')
@@ -17,18 +17,6 @@ def delete_connection_object(connection_id):
                 'connection_id': connection_id
             })
     return True
-
-
-def get_connection_id(event, context, body):
-    if context and hasattr(context, 'get') and context.get("connectionId"):
-        return context.get("connectionId")
-    if "connectionId" in event.get("requestContext", {}):
-        return event["requestContext"]["connectionId"]
-    if "connectionId" in event:
-        return event["connectionId"]
-    if "connectionId" in body:
-        return body["connectionId"]
-    raise ValueError("Request context is invalid!")
 
 
 def lambda_handler(event, context):
